+++ src/LambdaRobots/GameInfo.cs
@@ -36,39 +36,39 @@ namespace LambdaRobots {
         /// <summary>
         /// Width of the game board.
         /// </summary>
-        public double BoardWidth { get; set; }
+        public float BoardWidth { get; set; }
 
         /// <summary>
         /// Height of the game board.
         /// </summary>
-        public double BoardHeight { get; set; }
+        public float BoardHeight { get; set; }
 
         /// <summary>
         /// Number of seconds elapsed per game turn.
         /// </summary>
         ///
         // TODO: is still useful when using real-time?
-        public double SecondsPerTurn { get; set; }
+        public float SecondsPerTurn { get; set; }
 
         /// <summary>
         /// Distance for missile impact to count as direct hit.
         /// </summary>
-        public double DirectHitRange;
+        public float DirectHitRange;
 
         /// <summary>
         /// Distance for missile impact to count as near hit.
         /// </summary>
-        public double NearHitRange;
+        public float NearHitRange;
 
         /// <summary>
         /// Distance for missile impact to count as far hit.
         /// </summary>
-        public double FarHitRange;
+        public float FarHitRange;
 
         /// <summary>
         /// Distance between robots to count as a collision.
         /// </summary>
-        public double CollisionRange;
+        public float CollisionRange;
 
         /// <summary>
         /// Current game turn. Starts at `1`.
